:truck: Move expense type select to the expenses components.
