+++ web/src/components/expenses/EstimateCreateDialog.vue
@@ -100,7 +100,7 @@ import useTravelAuthorizationSummary from "@/use/travel-authorizations/use-trave
 
 import CurrencyTextField from "@/components/common/CurrencyTextField.vue"
 import StringDateInput from "@/components/common/StringDateInput.vue"
-import ExpenseTypeSelect from "@/modules/travel-authorizations/components/ExpenseTypeSelect.vue"
+import ExpenseTypeSelect from "@/components/expenses/ExpenseTypeSelect.vue"
 
 const emit = defineEmits(["created"])
