+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-expense-page/ExpenseCreateDialog.vue
@@ -96,7 +96,7 @@ import { type VForm } from "vuetify/components"
 import expensesApi, { ExpenseExpenseTypes, ExpenseTypes, type Expense } from "@/api/expenses-api"
 import CurrencyTextField from "@/components/common/CurrencyTextField.vue"
 import StringDateInput from "@/components/common/StringDateInput.vue"
-import ExpenseTypeSelect from "@/modules/travel-authorizations/components/ExpenseTypeSelect.vue"
+import ExpenseTypeSelect from "@/components/expenses/ExpenseTypeSelect.vue"
 import useSnack from "@/use/use-snack"
 import useRouteQuery, { booleanTransformer } from "@/use/utils/use-route-query"
 import { required } from "@/utils/validators"
